[Refactor] Simplify visitor method calls in AST statement classes
