+++ fe/fe-core/src/main/java/com/starrocks/service/FrontendServiceImpl.java
@@ -172,8 +172,6 @@ import com.starrocks.sql.ast.RangePartitionDesc;
 import com.starrocks.sql.ast.SetType;
 import com.starrocks.sql.ast.ShowAlterStmt;
 import com.starrocks.sql.ast.TableRef;
-import com.starrocks.sql.ast.TableRef;
-import com.starrocks.sql.parser.QualifiedName;
 import com.starrocks.sql.ast.expression.LiteralExpr;
 import com.starrocks.sql.common.StarRocksPlannerException;
 import com.starrocks.sql.parser.NodePosition;
