--- fe/fe-core/src/main/java/com/starrocks/sql/ast/DropTableStmt.java
@@ -1,91 +1,0 @@
-// Copyright 2021-present StarRocks, Inc. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-
-package com.starrocks.sql.ast;
-
-import com.starrocks.sql.parser.NodePosition;
-
-// DROP TABLE
-public class DropTableStmt extends DdlStmt {
-    private final boolean ifExists;
-    private TableRef tableRef;
-    private final boolean isView;
-    private final boolean forceDrop;
-
-    // used to mark whether it should be treated as a temporary table to distinguish subsequent processing logic,
-    // it should be set during analysis phase
-    private boolean temporaryTableMark = false;
-
-    public DropTableStmt(boolean ifExists, TableRef tableRef, boolean forceDrop) {
-        this(ifExists, tableRef, false, forceDrop, NodePosition.ZERO);
-    }
-
-    public DropTableStmt(boolean ifExists, TableRef tableRef, boolean isView, boolean forceDrop) {
-        this(ifExists, tableRef, isView, forceDrop, NodePosition.ZERO);
-    }
-
-    public DropTableStmt(boolean ifExists, TableRef tableRef, boolean isView, boolean forceDrop, NodePosition pos) {
-        super(pos);
-        this.ifExists = ifExists;
-        this.tableRef = tableRef;
-        this.isView = isView;
-        this.forceDrop = forceDrop;
-    }
-
-    public boolean isSetIfExists() {
-        return ifExists;
-    }
-
-    public TableRef getTableRef() {
-        return tableRef;
-    }
-
-    public void setTableRef(TableRef tableRef) {
-        this.tableRef = tableRef;
-    }
-
-    public String getCatalogName() {
-        return tableRef == null ? null : tableRef.getCatalogName();
-    }
-
-    public String getDbName() {
-        return tableRef == null ? null : tableRef.getDbName();
-    }
-
-    public String getTableName() {
-        return tableRef == null ? null : tableRef.getTableName();
-    }
-
-    public boolean isView() {
-        return isView;
-    }
-
-    public boolean isForceDrop() {
-        return this.forceDrop;
-    }
-
-    public void setTemporaryTableMark(boolean mark) {
-        this.temporaryTableMark = mark;
-    }
-
-    public boolean getTemporaryTableMark() {
-        return temporaryTableMark;
-    }
-
-    @Override
-    public <R, C> R accept(AstVisitor<R, C> visitor, C context) {
-        return ((AstVisitorExtendInterface<R, C>) visitor).visitDropTableStatement(this, context);
-    }
-}
