+++ fe/fe-parser/src/main/java/com/starrocks/sql/ast/AdminSetReplicaStatusStmt.java
@@ -70,6 +70,6 @@ public class AdminSetReplicaStatusStmt extends DdlStmt {
 
     @Override
     public <R, C> R accept(AstVisitor<R, C> visitor, C context) {
-        return ((AstVisitorExtendInterface<R, C>) visitor).visitAdminSetReplicaStatusStatement(this, context);
+        return visitor.visitAdminSetReplicaStatusStatement(this, context);
     }
 }
