+++ fe/fe-parser/src/main/java/com/starrocks/sql/ast/AnalyzeMultiColumnDesc.java
