+++ fe/fe-parser/src/main/java/com/starrocks/sql/ast/CancelRefreshMaterializedViewStmt.java
@@ -12,29 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-
 package com.starrocks.sql.ast;
 
-import com.starrocks.sql.ast.expression.Expr;
 import com.starrocks.sql.parser.NodePosition;
 
-import java.util.List;
-
-public class DropHistogramStmt extends StatementBase {
+public class CancelRefreshMaterializedViewStmt extends DdlStmt {
     private TableRef tableRef;
-    private List<String> columnNames;
-    private final List<Expr> columns;
+    private final boolean force;
 
-    private boolean isExternal = false;
-
-    public DropHistogramStmt(TableRef tableRef, List<Expr> columns) {
-        this(tableRef, columns, NodePosition.ZERO);
+    public CancelRefreshMaterializedViewStmt(TableRef tableRef, boolean force) {
+        this(tableRef, force, NodePosition.ZERO);
     }
 
-    public DropHistogramStmt(TableRef tableRef, List<Expr> columns, NodePosition pos) {
+    public CancelRefreshMaterializedViewStmt(TableRef tableRef, boolean force, NodePosition pos) {
         super(pos);
         this.tableRef = tableRef;
-        this.columns = columns;
+        this.force = force;
     }
 
     public TableRef getTableRef() {
@@ -53,32 +46,16 @@ public class DropHistogramStmt extends StatementBase {
         return tableRef == null ? null : tableRef.getDbName();
     }
 
-    public String getTableName() {
+    public String getMvName() {
         return tableRef == null ? null : tableRef.getTableName();
     }
 
-    public List<String> getColumnNames() {
-        return columnNames;
-    }
-
-    public void setColumnNames(List<String> columnNames) {
-        this.columnNames = columnNames;
-    }
-
-    public List<Expr> getColumns() {
-        return columns;
+    public boolean isForce() {
+        return force;
     }
 
     @Override
     public <R, C> R accept(AstVisitor<R, C> visitor, C context) {
-        return ((AstVisitorExtendInterface<R, C>) visitor).visitDropHistogramStatement(this, context);
-    }
-
-    public boolean isExternal() {
-        return isExternal;
-    }
-
-    public void setExternal(boolean isExternal) {
-        this.isExternal = isExternal;
+        return visitor.visitCancelRefreshMaterializedViewStatement(this, context);
     }
 }
