+++ fe/fe-parser/src/main/java/com/starrocks/sql/ast/RecoverTableStmt.java
@@ -51,6 +51,6 @@ public class RecoverTableStmt extends DdlStmt {
 
     @Override
     public <R, C> R accept(AstVisitor<R, C> visitor, C context) {
-        return ((AstVisitorExtendInterface<R, C>) visitor).visitRecoverTableStatement(this, context);
+        return visitor.visitRecoverTableStatement(this, context);
     }
 }
