+++ fe/fe-parser/src/main/java/com/starrocks/sql/ast/SetUserPropertyStmt.java
@@ -1,0 +1,52 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+
+package com.starrocks.sql.ast;
+
+import com.starrocks.sql.parser.NodePosition;
+
+import java.util.List;
+
+public class SetUserPropertyStmt extends DdlStmt {
+    private String user;
+    private final List<SetUserPropertyVar> propertyList;
+
+    public SetUserPropertyStmt(String user, List<SetUserPropertyVar> propertyList) {
+        this(user, propertyList, NodePosition.ZERO);
+    }
+
+    public SetUserPropertyStmt(String user, List<SetUserPropertyVar> propertyList, NodePosition pos) {
+        super(pos);
+        this.user = user;
+        this.propertyList = propertyList;
+    }
+
+    public String getUser() {
+        return user;
+    }
+
+    public void setUser(String user) {
+        this.user = user;
+    }
+
+    public List<SetUserPropertyVar> getPropertyList() {
+        return this.propertyList;
+    }
+
+    @Override
+    public <R, C> R accept(AstVisitor<R, C> visitor, C context) {
+        return visitor.visitSetUserPropertyStatement(this, context);
+    }
+}
