+++ fe/fe-parser/src/main/java/com/starrocks/sql/ast/ShowCreateTableStmt.java
@@ -72,6 +72,6 @@ public class ShowCreateTableStmt extends ShowStmt {
 
     @Override
     public <R, C> R accept(AstVisitor<R, C> visitor, C context) {
-        return ((AstVisitorExtendInterface<R, C>) visitor).visitShowCreateTableStatement(this, context);
+        return visitor.visitShowCreateTableStatement(this, context);
     }
 }
